+++ src/main/java/com/chaos/invoicify/repository/InvoicesRepository.java
@@ -1,5 +1,6 @@
 package com.chaos.invoicify.repository;
 
+import com.chaos.invoicify.entity.CompanyEntity;
 import com.chaos.invoicify.entity.InvoiceEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
