+++ src/main/java/com/chaos/invoicify/service/InvoiceService.java
@@ -34,28 +34,30 @@ public class InvoiceService {
             if (invoiceDto.getItemDtoList() != null) {
 
                 invoiceEntity.setItems(
-                    invoiceDto.getItemDtoList().stream()
-                        .map(
-                            itemDto -> new ItemEntity(
-                                itemDto.getItemDescription(),
-                                itemDto.getItemCount(),
-                                itemDto.getItemFeeType(),
-                                itemDto.getItemUnitPrice(),
-                                invoiceEntity)).collect(Collectors.toList()));
+                        invoiceDto.getItemDtoList().stream()
+                                .map(
+                                        itemDto -> new ItemEntity(
+                                                itemDto.getItemDescription(),
+                                                itemDto.getItemCount(),
+                                                itemDto.getItemFeeType(),
+                                                itemDto.getItemUnitPrice(),
+                                                invoiceEntity)).collect(Collectors.toList()));
             }
 
             invoicesRepository.save(invoiceEntity);
             newInvoiceDto = new InvoiceDto(invoiceEntity.getId(),
-                invoiceEntity.getCompany().getName(),
-                invoiceEntity.getCreateDate(),
-                invoiceEntity.getModifiedDate(),
-                invoiceEntity.getItems().stream()
-                    .map(
-                        itemEntity -> new ItemDto(
-                            itemEntity.getItemDescription(),
-                            itemEntity.getItemCount(),
-                            itemEntity.getItemFeeType(),
-                            itemEntity.getItemUnitPrice())).collect(Collectors.toList()));
+                    invoiceEntity.getCompany().getName(),
+                    invoiceEntity.getCreateDate(),
+                    invoiceEntity.getModifiedDate(),
+                    invoiceEntity.getItems().stream()
+                            .map(
+                                    itemEntity -> new ItemDto(
+                                            itemEntity.getItemDescription(),
+                                            itemEntity.getItemCount(),
+                                            itemEntity.getItemFeeType(),
+                                            itemEntity.getItemUnitPrice(),
+                                            itemEntity.getTotalItemValue()
+                                            )).collect(Collectors.toList()));
         }
 
         return newInvoiceDto;
@@ -63,20 +65,21 @@ public class InvoiceService {
 
     public List<InvoiceDto> fetchAllInvoices() {
         return invoicesRepository.findAll()
-            .stream()
-            .map(
-                invoiceEntity -> new InvoiceDto(
-                    invoiceEntity.getId(),
-                    invoiceEntity.getCompany().getName(),
-                    invoiceEntity.getCreateDate(),
-                    invoiceEntity.getModifiedDate(),
-                    invoiceEntity.getItems().stream()
-                        .map(itemEntity -> new ItemDto(
-                            itemEntity.getItemDescription(),
-                            itemEntity.getItemCount(),
-                            itemEntity.getItemFeeType(),
-                            itemEntity.getItemUnitPrice())).collect(Collectors.toList())
-                )).collect(Collectors.toList());
+                .stream()
+                .map(
+                        invoiceEntity -> new InvoiceDto(
+                                invoiceEntity.getId(),
+                                invoiceEntity.getCompany().getName(),
+                                invoiceEntity.getCreateDate(),
+                                invoiceEntity.getModifiedDate(),
+                                invoiceEntity.getItems().stream()
+                                        .map(itemEntity -> new ItemDto(
+                                                itemEntity.getItemDescription(),
+                                                itemEntity.getItemCount(),
+                                                itemEntity.getItemFeeType(),
+                                                itemEntity.getItemUnitPrice(),
+                                                itemEntity.getTotalItemValue())).collect(Collectors.toList())
+                        )).collect(Collectors.toList());
     }
 
     public InvoiceDto addItems(Long invoiceNumber, List<ItemDto> itemDtoList) {
@@ -85,46 +88,50 @@ public class InvoiceService {
 
         if (invoiceEntity != null) {
             itemDtoList
-                .forEach(
-                    itemDto -> {
-                        ItemEntity itemEntity =
-                            new ItemEntity(
-                                itemDto.getItemDescription(),
-                                itemDto.getItemCount(),
-                                itemDto.getItemFeeType(),
-                                itemDto.getItemUnitPrice(),
-                                invoiceEntity);
-                        this.itemRepository.save(itemEntity);
-                    });
+                    .forEach(
+                            itemDto -> {
+                                ItemEntity itemEntity =
+                                        new ItemEntity(
+                                                itemDto.getItemDescription(),
+                                                itemDto.getItemCount(),
+                                                itemDto.getItemFeeType(),
+                                                itemDto.getItemUnitPrice(),
+                                                invoiceEntity);
+                                this.itemRepository.save(itemEntity);
+                            });
             newInvoiceDto = new InvoiceDto(
-                invoiceEntity.getId(),
-                invoiceEntity.getCompany().getName(),
-                invoiceEntity.getCreateDate(),
-                invoiceEntity.getModifiedDate(),
-                invoiceEntity.getItems().stream()
-                    .map(itemEntity -> new ItemDto(
-                        itemEntity.getItemDescription(),
-                        itemEntity.getItemCount(),
-                        itemEntity.getItemFeeType(),
-                        itemEntity.getItemUnitPrice())).collect(Collectors.toList()));
+                    invoiceEntity.getId(),
+                    invoiceEntity.getCompany().getName(),
+                    invoiceEntity.getCreateDate(),
+                    invoiceEntity.getModifiedDate(),
+                    invoiceEntity.getItems().stream()
+                            .map(itemEntity -> new ItemDto(
+                                    itemEntity.getItemDescription(),
+                                    itemEntity.getItemCount(),
+                                    itemEntity.getItemFeeType(),
+                                    itemEntity.getItemUnitPrice(),
+                                    itemEntity.getTotalItemValue())).collect(Collectors.toList()));
         }
         return newInvoiceDto;
     }
 
     public List<InvoiceDto> fetchInvoiceById(Long invoiceNumber) {
         return invoicesRepository.findById(invoiceNumber)
-            .stream()
-            .map(invoiceEntity -> new InvoiceDto(
-                invoiceEntity.getId(),
-                invoiceEntity.getCompany().getName(),
-                invoiceEntity.getCreateDate(),
-                invoiceEntity.getModifiedDate(),
-                invoiceEntity.getItems().stream()
-                    .map(itemEntity -> new ItemDto(
-                        itemEntity.getItemDescription(),
-                        itemEntity.getItemCount(),
-                        itemEntity.getItemFeeType(),
-                        itemEntity.getItemUnitPrice())).collect(Collectors.toList())
-            )).collect(Collectors.toList());
+                .stream()
+                .map(invoiceEntity -> new InvoiceDto(
+                        invoiceEntity.getId(),
+                        invoiceEntity.getCompany().getName(),
+                        invoiceEntity.getCreateDate(),
+                        invoiceEntity.getModifiedDate(),
+                        invoiceEntity.getItems().stream()
+                                .map(itemEntity -> new ItemDto(
+                                        itemEntity.getItemDescription(),
+                                        itemEntity.getItemCount(),
+                                        itemEntity.getItemFeeType(),
+                                        itemEntity.getItemUnitPrice(),
+                                        itemEntity.getTotalItemValue())).collect(Collectors.toList())
+                )).collect(Collectors.toList());
     }
+
+
 }
